+++ src/OpenTelemetry.AutoInstrumentation/Instrumentation.cs
@@ -115,9 +115,10 @@ internal static class Instrumentation
                 {
                     var builder = Sdk
                         .CreateTracerProviderBuilder()
+                        .InvokePluginsBefore(_pluginManager)
                         .SetResourceBuilder(ResourceConfigurator.CreateResourceBuilder(GeneralSettings.Value.EnabledResourceDetectors))
                         .UseEnvironmentVariables(LazyInstrumentationLoader, TracerSettings.Value, _pluginManager)
-                        .InvokePlugins(_pluginManager);
+                        .InvokePluginsAfter(_pluginManager);
 
                     _tracerProvider = builder.Build();
                     Logger.Information("OpenTelemetry tracer initialized.");
@@ -135,9 +136,10 @@ internal static class Instrumentation
                 {
                     var builder = Sdk
                         .CreateMeterProviderBuilder()
+                        .InvokePluginsBefore(_pluginManager)
                         .SetResourceBuilder(ResourceConfigurator.CreateResourceBuilder(GeneralSettings.Value.EnabledResourceDetectors))
                         .UseEnvironmentVariables(LazyInstrumentationLoader, MetricSettings.Value, _pluginManager)
-                        .InvokePlugins(_pluginManager);
+                        .InvokePluginsAfter(_pluginManager);
 
                     _meterProvider = builder.Build();
                     Logger.Information("OpenTelemetry meter initialized.");