+++ CHANGELOG.md
@@ -9,7 +9,14 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
 
 ### Added
 
+- Added `BeforeConfigureTracerProvider` and `BeforeConfigureMeterProvider` for plugins.
+  See [plugins documentation](/docs/plugins.md) for details.
+
 ### Changed
+
+- In plugins `ConfigureTracerProvider` and `ConfigureMeterProvider` are changed now
+  to `AfterConfigureTracerProvider` and `AfterConfigureMeterProvider`.
+  See [plugins documentation](/docs/plugins.md) for details.
 
 ### Deprecated